src/workflows/training_workflow/requirements.dev.txt
wheel==0.46.3
pip==26.0.1
flytekit==1.16.14
flytekitplugins-ray==1.16.14
flytekitplugins-mlflow==1.16.14
ray==2.54.0
ray[data]==2.54.0
ray[train]==2.54.0
boto3==1.42.61
pyarrow==23.0.1
great_expectations==1.14.0
flaml==2.5.0
mlflow==3.10.1
lightgbm==4.6.0
scikit-learn==1.7.2
feast==0.60.0
tqdm==4.67.3
onnx==1.20.1

